learning/test3/my-pipecat-bot/pyproject.toml
[project]
name = "my-pipecat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pipecat-ai[all]>=0.0.94",
]
